api/requirements.txt
flask
python-dotenv
psycopg
flask-bcrypt
flask-session
flask-cors
flask-sqlalchemy
